Refactors install files
Pulls out common functions into new library.
